LegolandPython/requirements.txt
streamlit
pandas
snowflake-connector-python
plotly.express
datetime
pandasql
streamlit-aggrid
mysql-connector-python
sqlalchemy
plotly
requests